Updating styling
